requirements.txt
# Core Framework
streamlit>=1.28.0                    # Web interface framework for building the app UI
langchain>=0.1.0                     # Core library for Retrieval-Augmented Generation (RAG) and LLM integration
langchain-community>=0.0.10          # Community-contributed integrations for LangChain (e.g., loaders, retrievers)
langchain-core>=0.1.0               # Core abstractions and utilities for LangChain
langchain-groq>=0.0.1               # Groq LLM integration for fast language model inference
langchain-huggingface>=0.0.1        # HuggingFace integration for embeddings and model access
langchain-text-splitters>=0.0.1     # Utilities for splitting text into chunks for embedding

# Vector Database (FAISS instead of Chroma as requested)
faiss-cpu>=1.7.4                    # FAISS vector database for fast similarity search
numpy>=1.24.0                       # Required for FAISS and numerical operations

# Database
mysql-connector-python>=8.0.0       # MySQL database connector for storing chat history and feedback
pymysql>=1.1.0                      # Alternative MySQL connector for better compatibility

# Document Processing
PyPDF2>=3.0.0                       # PDF file reading and manipulation
pypdf>=3.15.0                       # Alternative PDF processing library

# Embeddings and ML
sentence-transformers>=2.2.0         # Provides embedding models (used by HuggingFaceEmbeddings)
torch>=2.0.0                        # Required for sentence-transformers

# Environment and Configuration
python-dotenv>=1.0.0                # Loads environment variables from .env files (for API keys)

# Additional Utilities
pandas>=2.0.0                       # Data manipulation for chat history and analytics
requests>=2.31.0                    # HTTP requests for API calls 